Regression testbenches now emit bits instead of ints

+++ synth_examples/jacobi2d_unrolled_1/regression_tb_jacobi2d_unrolled_1_opt.cpp
@@ -24,7 +24,7 @@ int main() {
   for (int i = 0; i < 256; i++) {
     hw_uint<32> actual = jacobi2d_unrolled_1_update_0_write.read();
     auto actual_lane_0 = actual.extract<0*32, 31>();
-    fout << (uint64_t) actual_lane_0 << endl;
+    fout << actual_lane_0 << endl;
   }
 
   in_pix.close();
